#158 - Add instructions button in wikifishlog screen
